+++ .gitignore
@@ -1,4 +1,4 @@
-nimcache
+/nimcache
 
 # Executables shall be put in an ignored build/ directory
 /build
