+++ nim.cfg
@@ -17,3 +17,7 @@
 --excessiveStackTrace:on
 -d:metrics # enable metric collection
 
+# Required to make up for the compiler's inability to invalidate the C file
+# (and object) cache on different command line arguments.
+--forceBuild
+
