--- nimbus/api/.gitignore
@@ -1,2 +1,0 @@
-status_api_test
-libnimbus_api.so
