--- nimbus/api/README.md
@@ -1,7 +1,0 @@
-This folder contains an experimental C API for using parts of the nimbus
-code from C/go in the status console client:
-
-https://github.com/status-im/status-console-client/
-
-It serves mainly as a proof-of-concept for now - there are several unresolved
-issues surrounding threading, inter-language communication, callbacks etc.
