--- nimbus/api/build_go.sh
@@ -1,2 +1,0 @@
-#!/bin/sh 
-go build --ldflags '-extldflags "-static"' -o test main.go
