--- nimbus/api/status_api.nim.cfg
@@ -1,4 +1,0 @@
--app:lib
-o:"libnimbus_api.so"
-noMain
--d:"chronicles_sinks=textlines"
