+++ wrappers/wrapper.nim
@@ -12,7 +12,7 @@ import
   nimcrypto/[bcmode, hmac, rijndael, pbkdf2, sha2, sysrand, utils, keccak, hash],
   eth/[keys, rlp, p2p], eth/p2p/rlpx_protocols/[whisper_protocol],
   eth/p2p/[discovery, enode, peer_pool], chronicles,
-  ../config
+  ../nimbus/config
 
 type
   CReceivedMessage* = object
